# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Uno
  version: "1"
  description: |
    Uno's HTTP surface: the builder API under `/api`, which the SPA calls, and
    the public Stremio-protocol addon under `/u/{token}`, which Nuvio calls.
    Everything else on the server is the embedded SPA and is not described
    here.

    **Errors.** Every error under `/api`, an unknown path included, is JSON
    (`Error`): `error` holds one line of words and, for an error the builder
    acts on by kind, `code` names it (`ProfileNotFound`). Push is the one
    exception: past auth, profile resolution and body decoding it answers
    `PushResult` whether it succeeds or fails. The addon under `/u/{token}`
    answers errors in plain text (`text/plain; charset=utf-8`), which its
    clients don't read.

    **Request bodies.** Every body is decoded strictly: a field the request
    doesn't have, at any depth, is a 400 naming it (`invalid request body:
    unknown field "name"`). The default limit is 1 MiB, past which the answer
    is 413.

    **Auth.** Every `/api` operation except `GET /api/health` and
    `GET /api/config` takes a Nuvio access token as `Authorization: Bearer`.
    Any of them can answer 401 (missing, malformed, invalid or expired
    token), 403 (an account the server's allowlist doesn't list) or 502
    (`auth service unavailable`: Nuvio's signing keys couldn't be fetched).
    Any operation can answer 500 for a failure of Uno's own.

...

...
    TMDB key or TMDB refuses it.

...

...
servers:
...
security:
  - nuvioBearer: []
paths:
...
  /api/account/tmdb-key:
    get:
      tags: [account]
      operationId: getTMDBKey
      summary: Whether the caller has saved a TMDB key
      responses:
        "200":
          description: The key's status; the key itself is never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TMDBKeyStatus"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
    put:
      tags: [account]
      operationId: putTMDBKey
      summary: Save the caller's TMDB key
      description: |
        Saves the key, replacing any saved one, once it has the shape of a
        TMDB API key and TMDB accepts it. Nothing is saved otherwise.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key:
                  type: string
                  description: A TMDB API Key (32 characters, 0–9 and a–f), not a Read Access Token.
      responses:
        "200":
          description: Saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TMDBKeyStatus"
        "400":
          description: The key isn't shaped like a TMDB API key, or TMDB refused it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "502": { $ref: "#/components/responses/TMDBUnavailable" }
    delete:
      tags: [account]
      operationId: deleteTMDBKey
      summary: Remove the caller's TMDB key
      responses:
        "204":
          description: Removed, or there was none.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SharedKeyMode" }
...
components:
  securitySchemes:
    nuvioBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Nuvio access token, verified against Nuvio's signing keys.
...
  responses:
...
    Unauthorized:
      description: Missing, malformed, invalid or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The server's allowlist doesn't list this Nuvio account.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    SharedKeyMode:
      description: The server shares one TMDB key; accounts have none of their own.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: The body is over its cap (1 MiB; 4 MiB for the import routes).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TMDBUnavailable:
      description: TMDB couldn't be reached, or the auth service is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    TMDBKeyStatus:
      type: object
      required: [set]
      properties:
        set:
          type: boolean
        last4:
          type: string
          description: The saved key's last four characters; absent when none is saved.
...
    Error:
      type: object
      description: The answer to every failed request under `/api` but push's own.
      required: [error]
      properties:
        error:
          type: string
          description: One line of words, for a person to read.
        code:
          type: string
          description: A stable name for an error the builder acts on by kind rather than by status; absent when there is none.
...